+++ tests/testthat/test-group.R
@@ -119,17 +119,17 @@ test_that("nnet::multinom: order of rows determined by formula terms", {
     trash <- capture.output(tab <- modelsummary(mod, "data.frame", group = response + term ~ model))
     expect_s3_class(tab, "data.frame")
     expect_equal(colnames(tab),
-                 c("part", "group", "term", "statistic", "Model 1", "Model 2"))
+                 c("part", "response", "term", "statistic", "Model 1", "Model 2"))
     expect_equal(tab$term[1:4], c("(Intercept)", "(Intercept)", "mpg", "mpg"))
-    expect_equal(tab$group[1:12], c(rep("6", 6), rep("8", 6)))
+    expect_equal(tab$response[1:12], c(rep("6", 6), rep("8", 6)))
 
     ## order of rows determined by order of formula terms
     trash <- capture.output(tab <- modelsummary(mod, "data.frame", group = term + response ~ model))
     expect_s3_class(tab, "data.frame")
     expect_equal(colnames(tab),
-                 c("part", "term", "group", "statistic", "Model 1", "Model 2"))
+                 c("part", "term", "response", "statistic", "Model 1", "Model 2"))
     expect_equal(tab$term[1:4], rep("(Intercept)", 4))
-    expect_equal(tab$group[1:4], c("6", "6", "8", "8"))
+    expect_equal(tab$response[1:4], c("6", "6", "8", "8"))
 
     ## order of rows determined by order of formula terms
     trash <- capture.output(tab <- modelsummary(mod, "data.frame", group = model + term ~ response))
@@ -158,7 +158,7 @@ test_that("group ~ model + term", {
                         output = "data.frame",
                         group = response ~ model + term,
                         metrics = "RMSE")
-    known <- c("part", "group", "statistic", "Model 1 / (Intercept)", "Model 1 / mpg",
+    known <- c("part", "response", "statistic", "Model 1 / (Intercept)", "Model 1 / mpg",
 "Model 2 / (Intercept)", "Model 2 / mpg", "Model 2 / drat")
     expect_equal(colnames(tab), known)
     expect_equal(dim(tab), c(6, 8))
@@ -167,7 +167,7 @@ test_that("group ~ model + term", {
                         output = "data.frame",
                         group = response ~ term + model,
                         metrics = "RMSE")
-    known <- c("part", "group", "statistic", "(Intercept) / Model 1", "(Intercept) / Model 2",
+    known <- c("part", "response", "statistic", "(Intercept) / Model 1", "(Intercept) / Model 2",
 "mpg / Model 1", "mpg / Model 2", "drat / Model 2")
     expect_equal(colnames(tab), known)
     expect_equal(dim(tab), c(6, 8))
@@ -226,12 +226,12 @@ test_that("grouped coefficients: gamlss", {
     tab <- modelsummary(mod, "data.frame", group = term + component ~ model)
     expect_s3_class(tab, "data.frame")
     expect_equal(colnames(tab),
-                 c("part", "term", "group", "statistic", "Model 1", "Model 2"))
+                 c("part", "term", "component", "statistic", "Model 1", "Model 2"))
 
     tab <- modelsummary(mod, "data.frame", group = component + term ~ model)
     expect_s3_class(tab, "data.frame")
     expect_equal(colnames(tab),
-                 c("part", "group", "term", "statistic", "Model 1", "Model 2"))
+                 c("part", "component", "term", "statistic", "Model 1", "Model 2"))
 
     tab <- modelsummary(mod, "data.frame", group = term ~ model + component)
     expect_s3_class(tab, "data.frame")