+++ tests/testthat/test-pkg-marginaleffects.R
@@ -23,3 +23,42 @@ test_that("no error", {
                               output = "data.frame"),
                  NA)
 })
+
+
+
+
+test_that("multiple groups", {
+    requiet("marginaleffects")
+
+    mod <- lm(mpg ~ am + vs + factor(cyl), data = mtcars)
+    mod2 <- lm(mpg ~ drat + am + vs + factor(cyl), data = mtcars)
+    cmp <- comparisons(mod, variables = c("am", "vs"), by = "cyl", cross = TRUE)
+    cmp2 <- comparisons(mod2, variables = c("am", "vs"), by = "cyl", cross = TRUE)
+
+    tab <- modelsummary(
+        list(cmp, cmp2),
+        output = "dataframe",
+        gof_map = NA,
+        shape = cyl + contrast_am + contrast_vs ~ model)
+    expect_equal(dim(tab), c(6, 7))
+
+    tab <- modelsummary(
+        cmp,
+        output = "dataframe",
+        gof_map = NA,
+        shape = contrast_vs + contrast_am + cyl ~ model)
+    expect_equal(dim(tab), c(6, 6))
+
+    tab <- modelsummary(
+        cmp,
+        output = "dataframe",
+        gof_map = NA,
+        shape = contrast_vs + contrast_am ~ cyl + model)
+    expect_equal(dim(tab), c(2, 7))
+
+    expect_error(
+        modelsummary(
+            cmp,
+            shape = term + cyl + trash + contrast_am + contrast_vs ~ model),
+        regexp = "not found")
+})